chore: migrate firestore module into the gcloud module
