--- docs/examples/firestore.md
@@ -1,9 +1,0 @@
-# Firestore
-
-<!--codeinclude-->
-[Creating a Firestore container](../../examples/firestore/firestore.go)
-<!--/codeinclude-->
-
-<!--codeinclude-->
-[Test for a Firestore container](../../examples/firestore/firestore_test.go)
-<!--/codeinclude-->
